Consolidate two components and rename for clarity
